Integ_aplicaciones_computacion/tarea3_cursor/requirements.txt
# AWS CDK Dependencies
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# Lambda Function Dependencies
boto3>=1.26.0
requests>=2.28.0

# Development Dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# AWS Lambda Powertools (for better Lambda development)
aws-lambda-powertools>=2.0.0

# Database connectors (for on-premise database connection)
pymysql>=1.0.0
psycopg2-binary>=2.9.0
pyodbc>=4.0.0

# Additional utilities
python-dotenv>=0.19.0
pydantic>=1.10.0
